+++ Cargo.toml
@@ -37,7 +37,7 @@ bytes = { version = "0.5", optional = true }
 futures-util = { version = "0.3.0", default-features = false, optional = true }
 pin-project-lite = { version = "0.1", optional = true }
 tokio-util = { version = "0.2", optional = true }
-tokio = { version = "0.2", optional = true }
+tokio = { version = "0.2.10", optional = true }
 
 # Only needed for the r2d2 feature
 r2d2 = { version = "0.8.8", optional = true }
